--- a/mlx/backend/metal/kernels/steel/gemm/kernels/steel_gemm_gather.h
+++ b/mlx/backend/metal/kernels/steel/gemm/kernels/steel_gemm_gather.h
@@ -19,8 +19,8 @@ template <
 [[kernel, max_total_threads_per_threadgroup(WM* WN * 32)]] void gather_mm_rhs(
     const device T* A [[buffer(0)]],
     const device T* B [[buffer(1)]],
-    device T* C [[buffer(2)]],
-    const device uint32_t* rhs_indices [[buffer(3)]],
+    const device uint32_t* rhs_indices [[buffer(2)]],
+    device T* C [[buffer(3)]],
     const constant GEMMParams* params [[buffer(4)]],
     uint simd_lane_id [[thread_index_in_simdgroup]],
     uint simd_group_id [[simdgroup_index_in_threadgroup]],
@@ -43,7 +43,8 @@ template <
   using loader_b_t = typename gemm_kernel::loader_b_t;
   using mma_t = typename gemm_kernel::mma_t;
 
-  if (params->tiles_n <= tid.x || params->tiles_m <= tid.y) {
+  if (params->tiles_n <= static_cast<int>(tid.x) ||
+      params->tiles_m <= static_cast<int>(tid.y)) {
     return;
   }
 
@@ -57,159 +58,189 @@ template <
   const size_t c_row_long = size_t(c_row);
   const size_t c_col_long = size_t(c_col);
 
-  A += transpose_a ? c_row_long : c_row_long * params->lda;
-  C += c_row_long * params->ldd + c_col_long;
-
-  // NOTE: This assumes that each thread reads from 1 row at most so this
-  //       should change and be in the block loader rather than in here.
-  constexpr int n_reads = BK * BN / (WM * WN * 32);
-  constexpr int threads_per_row = (transpose_b) ? BN / n_reads : BK / n_reads;
-  if (align_M) {
-    const int64_t offset = params->batch_stride_b *
-        (rhs_indices
-             [c_row + (simd_group_id * 32 + simd_lane_id) / threads_per_row]);
-    B += offset + (transpose_b ? c_col_long * params->ldb : c_col_long);
-  } else {
-    const int m = c_row + (simd_group_id * 32 + simd_lane_id) / threads_per_row;
-    const int64_t offset = params->batch_stride_b *
-        ((m < params->M) ? rhs_indices[m] : rhs_indices[params->M - 1]);
-    B += offset + (transpose_b ? c_col_long * params->ldb : c_col_long);
-  }
-
-  threadgroup_barrier(mem_flags::mem_none);
-
-  // Prepare threadgroup mma operation
-  thread mma_t mma_op(simd_group_id, simd_lane_id);
-
-  // Prepare threadgroup loading operations
-  thread loader_a_t loader_a(A, params->lda, As, simd_group_id, simd_lane_id);
-  thread loader_b_t loader_b(B, params->ldb, Bs, simd_group_id, simd_lane_id);
-
   // Prepare threadgroup bounds
   const short tgp_bm = align_M ? BM : short(min(BM, params->M - c_row));
   const short tgp_bn = align_N ? BN : short(min(BN, params->N - c_col));
 
-  // Prepare iterations
-  const int gemm_k_iterations = params->gemm_k_iterations_aligned;
+  A += transpose_a ? c_row_long : c_row_long * params->lda;
+  B += transpose_b ? c_col_long * params->ldb : c_col_long;
+  C += c_row_long * params->ldd + c_col_long;
 
-  // Do unaligned K iterations first
-  if (!align_K) {
-    const int k_last = params->gemm_k_iterations_aligned * BK;
-    const int k_remain = params->K - k_last;
-    const size_t k_jump_a =
-        transpose_a ? params->lda * size_t(k_last) : size_t(k_last);
-    const size_t k_jump_b =
-        transpose_b ? size_t(k_last) : params->ldb * size_t(k_last);
-
-    // Move loader source ahead to end
-    loader_a.src += k_jump_a;
-    loader_b.src += k_jump_b;
-
-    // Load tile
-    const short2 tile_dims_A =
-        transpose_a ? short2(tgp_bm, k_remain) : short2(k_remain, tgp_bm);
-    const short2 tile_dims_B =
-        transpose_b ? short2(k_remain, tgp_bn) : short2(tgp_bn, k_remain);
-
-    loader_a.load_safe(tile_dims_A);
-    loader_b.load_safe(tile_dims_B);
-
-    threadgroup_barrier(mem_flags::mem_threadgroup);
-
-    // Do matmul
-    mma_op.mma(As, Bs);
-
-    // Reset source back to start
-    loader_a.src -= k_jump_a;
-    loader_b.src -= k_jump_b;
+  // Calculate the number of unique matmuls
+  uint32_t unique_indices[4];
+  short offsets[5];
+  unique_indices[0] = rhs_indices[c_row];
+  offsets[0] = 0;
+  int j = 0;
+  for (int i = 1; i < tgp_bm; i++) {
+    if (rhs_indices[c_row + i] != unique_indices[j]) {
+      j++;
+      offsets[j] = i;
+      unique_indices[j] = rhs_indices[c_row + i];
+    }
   }
+  int n_unique = j + 1;
+  offsets[n_unique] = tgp_bm;
 
-  // Matrix level aligned never check
-  if (align_M && align_N) {
-    for (int k = 0; k < gemm_k_iterations; k++) {
+  threadgroup_barrier(mem_flags::mem_none);
+
+  for (int n = 0; n < n_unique; n++) {
+    // Prepare threadgroup mma operation
+    thread mma_t mma_op(simd_group_id, simd_lane_id);
+
+    // Prepare threadgroup loading operations
+    thread loader_a_t loader_a(A, params->lda, As, simd_group_id, simd_lane_id);
+    thread loader_b_t loader_b(
+        B + unique_indices[n] * params->batch_stride_b,
+        params->ldb,
+        Bs,
+        simd_group_id,
+        simd_lane_id);
+
+    // Prepare iterations
+    const int gemm_k_iterations = params->gemm_k_iterations_aligned;
+
+    // Do unaligned K iterations first
+    if (!align_K) {
+      const int k_last = params->gemm_k_iterations_aligned * BK;
+      const int k_remain = params->K - k_last;
+      const size_t k_jump_a =
+          transpose_a ? params->lda * size_t(k_last) : size_t(k_last);
+      const size_t k_jump_b =
+          transpose_b ? size_t(k_last) : params->ldb * size_t(k_last);
+
+      // Move loader source ahead to end
+      loader_a.src += k_jump_a;
+      loader_b.src += k_jump_b;
+
+      // Load tile
+      const short2 tile_dims_A =
+          transpose_a ? short2(tgp_bm, k_remain) : short2(k_remain, tgp_bm);
+      const short2 tile_dims_B =
+          transpose_b ? short2(k_remain, tgp_bn) : short2(tgp_bn, k_remain);
+
+      loader_a.load_safe(tile_dims_A);
+      loader_b.load_safe(tile_dims_B);
+
       threadgroup_barrier(mem_flags::mem_threadgroup);
 
-      // Load elements into threadgroup
-      loader_a.load_unsafe();
-      loader_b.load_unsafe();
-
-      threadgroup_barrier(mem_flags::mem_threadgroup);
-
-      // Multiply and accumulate threadgroup elements
+      // Do matmul
       mma_op.mma(As, Bs);
 
-      // Prepare for next iteration
-      loader_a.next();
-      loader_b.next();
+      // Reset source back to start
+      loader_a.src -= k_jump_a;
+      loader_b.src -= k_jump_b;
     }
 
-    // Store results to device memory
-    return mma_op.store_result(C, params->ldd);
-  } else {
-    const short lbk = 0;
+    // Matrix level aligned never check
+    if (align_M && align_N) {
+      for (int k = 0; k < gemm_k_iterations; k++) {
+        threadgroup_barrier(mem_flags::mem_threadgroup);
 
-    // Tile aligned don't check
-    if ((align_M || tgp_bm == BM) && (align_N || tgp_bn == BN)) {
-      gemm_kernel::gemm_loop(
-          As,
-          Bs,
-          gemm_k_iterations,
-          loader_a,
-          loader_b,
-          mma_op,
-          tgp_bm,
-          tgp_bn,
-          lbk,
-          LoopAlignment<true, true, true>{});
-      return mma_op.store_result(C, params->ldd);
-    }
+        // Load elements into threadgroup
+        loader_a.load_unsafe();
+        loader_b.load_unsafe();
 
-    // Tile partially aligned check rows
-    else if (align_N || tgp_bn == BN) {
-      gemm_kernel::gemm_loop(
-          As,
-          Bs,
-          gemm_k_iterations,
-          loader_a,
-          loader_b,
-          mma_op,
-          tgp_bm,
-          tgp_bn,
-          lbk,
-          LoopAlignment<false, true, true>{});
-      return mma_op.store_result(C, params->ldd);
-    }
+        threadgroup_barrier(mem_flags::mem_threadgroup);
 
-    // Tile partially aligned check cols
-    else if (align_M || tgp_bm == BM) {
-      gemm_kernel::gemm_loop(
-          As,
-          Bs,
-          gemm_k_iterations,
-          loader_a,
-          loader_b,
-          mma_op,
-          tgp_bm,
-          tgp_bn,
-          lbk,
-          LoopAlignment<true, false, true>{});
-      return mma_op.store_result(C, params->ldd);
-    }
+        // Multiply and accumulate threadgroup elements
+        mma_op.mma(As, Bs);
 
-    // Nothing aligned so check both rows and cols
-    else {
-      gemm_kernel::gemm_loop(
-          As,
-          Bs,
-          gemm_k_iterations,
-          loader_a,
-          loader_b,
-          mma_op,
-          tgp_bm,
-          tgp_bn,
-          lbk,
-          LoopAlignment<false, false, true>{});
-      return mma_op.store_result(C, params->ldd);
+        // Prepare for next iteration
+        loader_a.next();
+        loader_b.next();
+      }
+
+      // Store results to device memory
+      if (offsets[n + 1] - offsets[n] == BM) {
+        mma_op.store_result(C, params->ldd);
+      } else {
+        mma_op.store_result_slice(
+            C, params->ldd, short2(0, offsets[n]), short2(BN, offsets[n + 1]));
+      }
+    } else {
+      const short lbk = 0;
+
+      // Tile aligned don't check
+      if ((align_M || tgp_bm == BM) && (align_N || tgp_bn == BN)) {
+        gemm_kernel::gemm_loop(
+            As,
+            Bs,
+            gemm_k_iterations,
+            loader_a,
+            loader_b,
+            mma_op,
+            tgp_bm,
+            tgp_bn,
+            lbk,
+            LoopAlignment<true, true, true>{});
+        if (offsets[n + 1] - offsets[n] == BM) {
+          mma_op.store_result(C, params->ldd);
+        } else {
+          mma_op.store_result_slice(
+              C,
+              params->ldd,
+              short2(0, offsets[n]),
+              short2(BN, offsets[n + 1]));
+        }
+      }
+
+      // Tile partially aligned check rows
+      else if (align_N || tgp_bn == BN) {
+        gemm_kernel::gemm_loop(
+            As,
+            Bs,
+            gemm_k_iterations,
+            loader_a,
+            loader_b,
+            mma_op,
+            tgp_bm,
+            tgp_bn,
+            lbk,
+            LoopAlignment<false, true, true>{});
+        mma_op.store_result_slice(
+            C, params->ldd, short2(0, offsets[n]), short2(BN, offsets[n + 1]));
+      }
+
+      // Tile partially aligned check cols
+      else if (align_M || tgp_bm == BM) {
+        gemm_kernel::gemm_loop(
+            As,
+            Bs,
+            gemm_k_iterations,
+            loader_a,
+            loader_b,
+            mma_op,
+            tgp_bm,
+            tgp_bn,
+            lbk,
+            LoopAlignment<true, false, true>{});
+        mma_op.store_result_slice(
+            C,
+            params->ldd,
+            short2(0, offsets[n]),
+            short2(tgp_bn, offsets[n + 1]));
+      }
+
+      // Nothing aligned so check both rows and cols
+      else {
+        gemm_kernel::gemm_loop(
+            As,
+            Bs,
+            gemm_k_iterations,
+            loader_a,
+            loader_b,
+            mma_op,
+            tgp_bm,
+            tgp_bn,
+            lbk,
+            LoopAlignment<false, false, true>{});
+        mma_op.store_result_slice(
+            C,
+            params->ldd,
+            short2(0, offsets[n]),
+            short2(tgp_bn, offsets[n + 1]));
+      }
     }
   }
 }