… benchmarks/python/gather_mm_bench.py
… mlx/backend/common/CMakeLists.txt
… mlx/backend/metal/CMakeLists.txt
… mlx/backend/metal/kernels/CMakeLists.txt
… mlx/backend/metal/kernels/steel/gemm/kernels/steel_gemm_gather.metal
… mlx/backend/metal/kernels/steel/gemm/mma.h
… mlx/backend/metal/nojit_kernels.cpp
… mlx/backend/metal/utils.h
… mlx/ops.cpp
… mlx/ops.h
… mlx/primitives.cpp
… mlx/primitives.h
… python/src/ops.cpp
… python/tests/test_blas.py