This is probably good enough to start reviewing. I need to fix the JIT build and also add some benchmark numbers. For now I am only routing matrix vector products but I think small matrices could also be routed there.

So the change list is

- Move gather out of `steel_gemm_fused` so that both are easier to edit
- Refactor GatherMM::eval_gpu to just route so that it is a bit easier to follow and add specializations etc
- Add `store_slice` to `BlockMMA` and the tiles internally so we can store only a slice of the output (a variation of `store_safe` basically)

There is more to be done in refactoring steel and simplifying routing and computing the arguments for matmul, I haven't done much there at all. I am sure Jagrit agrees and has a lot of stuff cooking.