@angeloskath 2025-04-05
I am running the benchmark in the repo with different parameters and reporting below

```
N = batch size * seq len
D = model dims
M = hidden dims
E = number of experts
I = active experts
```

On an M2 Ultra

N | D | M | E | I | Before unsorted (ms) | Before sorted (ms) | Now sorted (ms)
--- | --- | --- | --- | --- | --- | --- | --- 
1024 | 1024 | 1024 | 32 | 4 | 12.7 | 10.2 | 1.9
1024 | 1024 | 1024 | 8 | 4 | 10.3 | 10.2 | 1.7
1024 | 4096 | 1024 | 256 | 4 | 171.8 | 40.6 | 26.6
1024 | 4096 | 1024 | 256 | 8 | 345 | 79 | 35.3
1024 | 4096 | 1024 | 8 | 4 | 46.9 | 39.4 | 5.5
512 | 1024 | 1024 | 256 | 4 | 19 | 5.8 | 4.8
512 | 1024 | 1024 | 256 | 1 | 5.3 | 2.8 | 3.4 

On my laptop

N | D | M | E | I | Before unsorted (ms) | Before sorted (ms) | Now sorted (ms)
--- | --- | --- | --- | --- | --- | --- | --- 
1024 | 1024 | 1024 | 32 | 4 | 268 | 68.1 | 10.8
1024 | 1024 | 1024 | 8 | 4 | 149 | 68.1 | 9.6
512 | 1024 | 1024 | 256 | 4 | 190 | 35 | 37
512 | 1024 | 1024 | 256 | 1 | 48 | 23 | 26

Basically if there are enough samples to fill the experts we see very significant speedups otherwise we can even have some slowdown. The slowdown is in cases where we are accessing many experts in each group of 16 inputs. Since in the benchmark the experts are also uniformly distributed, it is a kind of worst case scenario.